src/core/orchestrator/solver/python_interpreter-e2b/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
e2b-code-interpreter==1.5.2
boto3==1.35.0